api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.5
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.1
alembic==1.13.2
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-dateutil==2.9.0.post0
aiofiles==24.1.0
requests==2.32.3
starlette==0.38.5
jinja2==3.1.4
swagger-ui-bundle==0.0.9



